Applying Naive Bayes to the Data
Starting with, we loaded, pre-processed, and then trained data onto the Logistic Regression Model. 
Now to test its accuracy we are applying a naive Bayes algorithm and hence now if you test any tweet/sentence of your own, it classifies as positive or negative depending on the intent of the sentence.